Button styling is now handled in a separate method, allowing both player button instances to use the same code to improve code efficiency and remove redundant lines.
checkkRules now takes the player and their opponents associated data as parameters, so there is no need to have separate executions for both players.

Fixed a bug in checkDuplicates where if there was no duplicates, the method would return 0.

Added the method to check if the game had ended and started one to calculate the winner of the game.